requirements.txt
streamlit
openai
trafilatura
PyPDF2
tavily-python
fpdf
python-dotenv
